+++ src/api.ts
@@ -1,6 +1,5 @@
-import { Person } from './types/Person';
+import { Person } from './types';
 
-// eslint-disable-next-line operator-linebreak
 const API_URL =
   'https://mate-academy.github.io/react_people-table/api/people.json';
 
@@ -9,7 +8,6 @@ function wait(delay: number) {
 }
 
 export async function getPeople(): Promise<Person[]> {
-  // keep this delay for testing purpose
   return wait(500)
     .then(() => fetch(API_URL))
     .then(response => response.json());
