+++ src/components/ParentLink.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import classNames from 'classnames';
 import { Link, useSearchParams } from 'react-router-dom';
-import { Person } from '../types/Person';
+import { Person } from '../types';
 
 type Props = {
   person: Person;
