+++ src/utils/sortHelper.ts
@@ -1,5 +1,4 @@
-import { Person } from '../types/Person';
-import { Params, SortColumns } from '../types/Filter';
+import { Person, Params, SortColumns } from '../types';
 
 export const filterPeople = (people: Person[], params: Params): Person[] => {
   const { sortBy, sortOrder, query, sex, centuries } = params;
@@ -10,75 +9,40 @@ export const filterPeople = (people: Person[], params: Params): Person[] => {
 
   const filteredPeople = people
     .filter(person => person.name.toLowerCase().includes(query.toLowerCase()))
-    .filter(person => {
-      if (!sex) {
-        return true;
-      }
-
-      return person.sex === sex;
-    })
-    .filter(person => {
-      if (!centuries.length) {
-        return true;
-      }
-
-      return centuries.includes(getPersonCentury(person));
-    })
+    .filter(person => !sex || person.sex === sex)
+    .filter(
+      person =>
+        !centuries.length || centuries.includes(getPersonCentury(person)),
+    )
     .toSorted((person1, person2) => {
       switch (sortBy) {
         case SortColumns.Born: {
-          if (sortOrder === 'asc') {
-            return person1.born - person2.born;
-          }
-
-          if (sortOrder === 'desc') {
-            return person2.born - person1.born;
-          }
-
-          break;
+          return sortOrder === 'asc'
+            ? person1.born - person2.born
+            : person2.born - person1.born;
         }
 
         case SortColumns.Died: {
-          if (sortOrder === 'asc') {
-            return person1.died - person2.died;
-          }
-
-          if (sortOrder === 'desc') {
-            return person2.died - person1.died;
-          }
-
-          break;
+          return sortOrder === 'asc'
+            ? person1.died - person2.died
+            : person2.died - person1.died;
         }
 
         case SortColumns.Name: {
-          if (sortOrder === 'asc') {
-            return person1.name.localeCompare(person2.name);
-          }
-
-          if (sortOrder === 'desc') {
-            return person2.name.localeCompare(person1.name);
-          }
-
-          break;
+          return sortOrder === 'asc'
+            ? person1.name.localeCompare(person2.name)
+            : person2.name.localeCompare(person1.name);
         }
 
         case SortColumns.Sex: {
-          if (sortOrder === 'asc') {
-            return person1.sex.localeCompare(person2.sex);
-          }
-
-          if (sortOrder === 'desc') {
-            return person2.sex.localeCompare(person1.sex);
-          }
-
-          break;
+          return sortOrder === 'asc'
+            ? person1.sex.localeCompare(person2.sex)
+            : person2.sex.localeCompare(person1.sex);
         }
 
         default:
           return 0;
       }
-
-      return 0;
     });
 
   return filteredPeople;
